+++ system/src/Grav/Framework/Flex/FlexObject.php
@@ -14,6 +14,7 @@ use Grav\Common\Debugger;
 use Grav\Common\Grav;
 use Grav\Common\Inflector;
 use Grav\Common\Twig\Twig;
+use Grav\Common\User\Interfaces\UserInterface;
 use Grav\Common\Utils;
 use Grav\Framework\Cache\CacheInterface;
 use Grav\Framework\ContentBlock\HtmlBlock;
@@ -1022,5 +1023,38 @@ class FlexObject implements FlexObjectInterface, FlexAuthorizeInterface
     protected function createFormObject(string $name, array $options = null)
     {
         return new FlexForm($name, $this, $options);
+    }
+
+    /**
+     * @param string $action
+     * @return string
+     */
+    protected function getAuthorizeAction(string $action): string
+    {
+        // Handle special action save, which can mean either update or create.
+        if ($action === 'save') {
+            $action = $this->exists() ? 'update' : 'create';
+        }
+
+        return $action;
+    }
+
+    /**
+     * @return UserInterface|null
+     */
+    protected function getActiveUser(): ?UserInterface
+    {
+        /** @var UserInterface|null $user */
+        $user = Grav::instance()['user'] ?? null;
+
+        return $user;
+    }
+
+    /**
+     * @return string
+     */
+    protected function getAuthorizeScope(): string
+    {
+        return isset(Grav::instance()['admin']) ? 'admin' : 'site';
     }
 }